fix: rewrote agreement service

+++ service-user/src/main/java/by/salary/serviceuser/service/OrganisationService.java
@@ -3,6 +3,8 @@ package by.salary.serviceuser.service;
 import by.salary.serviceuser.entities.Organisation;
 import by.salary.serviceuser.entities.Permission;
 import by.salary.serviceuser.entities.User;
+import by.salary.serviceuser.exceptions.OrganisationNotFoundException;
+import by.salary.serviceuser.exceptions.UserNotFoundException;
 import by.salary.serviceuser.model.OrganisationRequestDTO;
 import by.salary.serviceuser.model.OrganisationResponseDTO;
 import by.salary.serviceuser.model.UserResponseDTO;
@@ -12,6 +14,7 @@ import by.salary.serviceuser.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.weaver.ast.Or;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
@@ -19,6 +22,7 @@ import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -37,12 +41,10 @@ public class OrganisationService {
         this.webClientBuilder = webClientBuilder;
         this.organisationRepository = organisationRepository;
         this.userRepository = userRepository;
-
     }
 
 
     public List<OrganisationResponseDTO> getAllOrganisations() {
-
         List<OrganisationResponseDTO> organisations = new ArrayList<>();
         organisationRepository.findAll().forEach(organisation -> {
             organisations.add(new OrganisationResponseDTO(organisation));
@@ -52,7 +54,11 @@ public class OrganisationService {
 
 
     public OrganisationResponseDTO getOneOrganisation(BigInteger id) {
-        return new OrganisationResponseDTO(organisationRepository.findById(id).get());
+        Optional<Organisation> optionalOrganisation = organisationRepository.findById(id);
+        if (optionalOrganisation.isEmpty()) {
+            throw new OrganisationNotFoundException("Organisation with id " + id + " not found", HttpStatus.NOT_FOUND);
+        }
+        return new OrganisationResponseDTO(optionalOrganisation.get());
     }
 
     private BigInteger createNewAgreementId() {
@@ -68,31 +74,49 @@ public class OrganisationService {
 
     public OrganisationResponseDTO createOrganisation(OrganisationRequestDTO organisationRequestDTO,
                                                       String email) {
-        User director = userRepository.findByUserEmail(email).get();
-        Organisation organisation = new Organisation(organisationRequestDTO, director);
+        Optional<User> directorOpt = userRepository.findByUserEmail(email);
+        if(directorOpt.isEmpty()){
+            throw new UserNotFoundException("User with email " + email + " not found", HttpStatus.NOT_FOUND);
+        }
+        User director = directorOpt.get();
+
+        Organisation organisation = new Organisation(organisationRequestDTO);
+        organisation.setAgreementId(createNewAgreementId());
+        organisationRepository.save(organisation);
+
         director.setOrganisation(organisation);
+
+        organisation.setDirector(director);
+
         director.setUserFirstName(organisationRequestDTO.getDirectorFirstName());
         director.setUserSurname(organisationRequestDTO.getDirectorSurname());
         director.setUserSecondName(organisationRequestDTO.getDirectorSecondName());
 
-        //grant all permissions to Director
-        director.addPermission(permissionRepository.findByName("*").get());
+        //TODO: add permissions
+        //director.addPermission(permissionRepository.findByName("*").get());
 
         userRepository.save(director);
-        organisation.setAgreementId(createNewAgreementId());
-        return new OrganisationResponseDTO(organisationRepository.save(organisation));
+
+        return new OrganisationResponseDTO(organisation);
     }
 
 
 
     public void deleteOrganisation(BigInteger id) {
+        if(!organisationRepository.existsById(id)){
+            throw new OrganisationNotFoundException("Organisation with id " + id + " not found", HttpStatus.NOT_FOUND);
+        }
         organisationRepository.deleteById(id);
     }
 
 
     public List<UserResponseDTO> getOrganisationUsers(BigInteger id) {
+        Optional<Organisation> optionalOrganisation = organisationRepository.findById(id);
+        if (optionalOrganisation.isEmpty()) {
+            throw new OrganisationNotFoundException("Organisation with id " + id + " not found", HttpStatus.NOT_FOUND);
+        }
         List<UserResponseDTO> users = new ArrayList<>();
-        organisationRepository.findById(id).get().getUsers().forEach(user -> {
+        optionalOrganisation.get().getUsers().forEach(user -> {
             users.add(new UserResponseDTO(user));
         });
         return users;
